refactor: move files to modules, extract persistence
