+++ src-tauri/src/state/mod.rs
@@ -1,0 +1,3 @@
+pub mod state;
+pub mod reducer;
+pub mod persistence;
